stuck at array
