+++ starwars/src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.css';
-import Characters from './index.js'
+import Characters from './components/Character.js'
 
 const App = () => {
   // Try to think through what state you'll need for this app before starting. Then build out
