+++ starwars/src/components/starWarsCard.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+const starWarsCard = props => {
+  return (
+    <div className="charcter-list">
+    </div>
+  );
+};
+
+export default starWarsCard;
